+++ src/wattwatt/interfaces/electricmeter/IElectricMeter.java
@@ -5,5 +5,5 @@ import fr.sorbonne_u.components.interfaces.RequiredI;
 
 public interface IElectricMeter extends OfferedI, RequiredI{
 
-	public int getAllConso() throws Exception;
+	public double getAllConso() throws Exception;
 }
